Move cursors t tools/cursors. Exposed these cursors to the lib, and added a getter to retrieve the SVG icon from MouseCursors.
