+++ src/lib.js
@@ -18,6 +18,7 @@ import {
 } from './util/findAndMoveHelpers.js';
 
 import mixins from './mixins/index.js';
+import cursors from './tools/cursors/index.js';
 
 import {
   getNewContext,
@@ -66,7 +67,7 @@ import {
 import angleBetweenPoints from './util/angleBetweenPoints.js';
 import getKeyFromKeyCode from './util/getKeyFromKeyCode.js';
 import numbersWithCommas from './util/numbersWithCommas.js';
-import MouseCursor from './util/MouseCursor.js';
+import MouseCursor from './tools/cursors/MouseCursor.js';
 
 import ellipseUtils from './util/ellipse/index.js';
 import freehandUtils from './util/freehand/index.js';
@@ -79,6 +80,9 @@ export const lib = {
   'base/BaseTool': BaseTool,
   'base/BaseAnnotationTool': BaseAnnotationTool,
   'base/BaseBrushTool': BaseBrushTool,
+
+  'tools/cursors/MouseCursor': MouseCursor,
+  'tools/cursors': cursors,
 
   'manipulators/anyHandlesOutsideImage': anyHandlesOutsideImage,
   'manipulators/getHandleNearImagePoint': getHandleNearImagePoint,
@@ -136,7 +140,6 @@ export const lib = {
   'util/lineSegDistance': lineSegDistance,
   'util/triggerEvent': triggerEvent,
   'util/convertToVectro3': convertToVector3,
-  'util/MouseCursor': MouseCursor,
 
   // Whole tool specific util packages
   'util/ellipseUtils': ellipseUtils,
