+++ src/tools/CrosshairsTool.js
@@ -11,6 +11,7 @@ import toolColors from '../stateManagement/toolColors.js';
 import { imagePointToPatientPoint } from '../util/pointProjector.js';
 import convertToVector3 from '../util/convertToVector3.js';
 import { setToolOptions } from '../toolOptions.js';
+import { crosshairsCursor } from './cursors/index.js';
 
 /**
  * @public
@@ -192,24 +193,3 @@ export default class CrosshairsTool extends BaseTool {
     });
   }
 }
-
-const crosshairsCursor = {
-  svgGroupString: `<path fill="ACTIVE_COLOR" d="M1325 1024h-109q-26 0-45-19t-19-45v-128q0-26
-    19-45t45-19h109q-32-108-112.5-188.5t-188.5-112.5v109q0 26-19 45t-45
-    19h-128q-26 0-45-19t-19-45v-109q-108 32-188.5 112.5t-112.5 188.5h109q26
-    0 45 19t19 45v128q0 26-19 45t-45 19h-109q32 108 112.5 188.5t188.5
-    112.5v-109q0-26 19-45t45-19h128q26 0 45 19t19 45v109q108-32
-    188.5-112.5t112.5-188.5zm339-192v128q0 26-19 45t-45 19h-143q-37 161-154.5
-    278.5t-278.5 154.5v143q0 26-19 45t-45 19h-128q-26
-    0-45-19t-19-45v-143q-161-37-278.5-154.5t-154.5-278.5h-143q-26
-    0-45-19t-19-45v-128q0-26 19-45t45-19h143q37-161
-    154.5-278.5t278.5-154.5v-143q0-26 19-45t45-19h128q26 0 45 19t19 45v143q161
-    37 278.5 154.5t154.5 278.5h143q26 0 45 19t19 45z"
-  />`,
-  options: {
-    viewBox: {
-      x: 1792,
-      y: 1792,
-    },
-  },
-};
