+++ src/tools/EraserTool.js
@@ -2,6 +2,7 @@ import external from '../externalModules.js';
 import BaseTool from './base/BaseTool.js';
 import { getToolState, removeToolState } from '../stateManagement/toolState.js';
 import { state } from '../store/index.js';
+import { eraserCursor } from './cursors/index.js';
 
 /**
  * @public
@@ -54,17 +55,3 @@ export default class EraserTool extends BaseTool {
     return consumeEvent;
   }
 }
-
-const eraserCursor = {
-  svgGroupString: `<path transform="translate(0,1792) scale(1,-1)" fill="ACTIVE_COLOR" d="M960 1408l336-384h-768l-336 384h768zm1013-1077q15
-      34 9.5 71.5t-30.5 65.5l-896 1024q-38 44-96 44h-768q-38
-      0-69.5-20.5t-47.5-54.5q-15-34-9.5-71.5t30.5-65.5l896-1024q38-44 96-44h768q38
-      0 69.5 20.5t47.5 54.5z"
-    />`,
-  options: {
-    viewBox: {
-      x: 2048,
-      y: 1792,
-    },
-  },
-};
