+++ src/tools/PanTool.js
@@ -1,5 +1,6 @@
 import external from './../externalModules.js';
 import BaseTool from './base/BaseTool.js';
+import { panCursor } from './cursors/index.js';
 
 /**
  * @public
@@ -60,20 +61,3 @@ export default class PanTool extends BaseTool {
     viewport.translation.y += translation.y;
   }
 }
-
-const panCursor = {
-  svgGroupString: `<path fill="ACTIVE_COLOR" d="M1411 541l-355 355 355 355 144-144q29-31 70-14 39 17
-      39 59v448q0 26-19 45t-45 19h-448q-42 0-59-40-17-39 14-69l144-144-355-355-355
-      355 144 144q31 30 14 69-17 40-59 40h-448q-26 0-45-19t-19-45v-448q0-42 40-59
-      39-17 69 14l144 144 355-355-355-355-144 144q-19 19-45 19-12
-      0-24-5-40-17-40-59v-448q0-26 19-45t45-19h448q42 0 59 40 17 39-14 69l-144
-      144 355 355 355-355-144-144q-31-30-14-69 17-40 59-40h448q26 0 45 19t19
-      45v448q0 42-39 59-13 5-25 5-26 0-45-19z"
-    />`,
-  options: {
-    viewBox: {
-      x: 1792,
-      y: 1792,
-    },
-  },
-};
