+++ src/tools/RotateTool.js
@@ -1,6 +1,7 @@
 import external from './../externalModules.js';
 import BaseTool from './base/BaseTool.js';
 import angleBetweenPoints from '../util/angleBetweenPoints.js';
+import { rotateCursor } from './cursors/index.js';
 
 /**
  * @public
@@ -98,21 +99,4 @@ const verticalStrategy = evt => {
   const { viewport, deltaPoints } = eventData;
 
   viewport.rotation += deltaPoints.page.y / viewport.scale;
-};
-
-const rotateCursor = {
-  svgGroupString: `<path fill="ACTIVE_COLOR" d="M1664 256v448q0 26-19 45t-45 19h-448q-42 0-59-40-17-39
-      14-69l138-138q-148-137-349-137-104 0-198.5 40.5t-163.5 109.5-109.5
-      163.5-40.5 198.5 40.5 198.5 109.5 163.5 163.5 109.5 198.5 40.5q119 0
-      225-52t179-147q7-10 23-12 15 0 25 9l137 138q9 8 9.5 20.5t-7.5 22.5q-109
-      132-264 204.5t-327 72.5q-156 0-298-61t-245-164-164-245-61-298 61-298
-      164-245 245-164 298-61q147 0 284.5 55.5t244.5 156.5l130-129q29-31 70-14
-      39 17 39 59z"
-    />`,
-  options: {
-    viewBox: {
-      x: 1792,
-      y: 1792,
-    },
-  },
 };
