+++ src/tools/annotation/AngleTool.js
@@ -21,6 +21,7 @@ import { textBoxWidth } from './../../drawing/drawTextBox.js';
 import drawHandles from './../../drawing/drawHandles.js';
 import lineSegDistance from './../../util/lineSegDistance.js';
 import roundToDecimal from './../../util/roundToDecimal.js';
+import { angleCursor } from '../cursors/index.js';
 
 /**
  * @public
@@ -32,7 +33,7 @@ import roundToDecimal from './../../util/roundToDecimal.js';
  *
  * @param {ToolConfiguration} [configuration={}]
  */
-class AngleTool extends BaseAnnotationTool {
+export default class AngleTool extends BaseAnnotationTool {
   constructor(configuration = {}) {
     const defaultConfig = {
       name: 'Angle',
@@ -327,18 +328,3 @@ class AngleTool extends BaseAnnotationTool {
 function length(vector) {
   return Math.sqrt(Math.pow(vector.x, 2) + Math.pow(vector.y, 2));
 }
-
-export default AngleTool;
-
-const angleCursor = {
-  svgGroupString: `<path fill="ACTIVE_COLOR" d="M1203 544q0 13-10 23l-393 393 393 393q10 10 10 23t-10 23l-50
-      50q-10 10-23 10t-23-10l-466-466q-10-10-10-23t10-23l466-466q10-10 23-10t23
-      10l50 50q10 10 10 23z"
-    />`,
-  options: {
-    viewBox: {
-      x: 1792,
-      y: 1792,
-    },
-  },
-};
