+++ src/tools/annotation/ProbeTool.js
@@ -11,6 +11,7 @@ import drawHandles from './../../drawing/drawHandles.js';
 // Utilities
 import getRGBPixels from './../../util/getRGBPixels.js';
 import calculateSUV from './../../util/calculateSUV.js';
+import { probeCursor } from '../cursors/index.js';
 
 /**
  * @public
@@ -180,18 +181,3 @@ export default class ProbeTool extends BaseAnnotationTool {
     }
   }
 }
-
-const probeCursor = {
-  svgGroupString: `<path fill="ACTIVE_COLOR" d="M1152 896q0 106-75 181t-181 75-181-75-75-181 75-181 181-75 181 75
-    75 181zm-256-544q-148 0-273 73t-198 198-73 273 73 273 198 198 273 73 273-73
-    198-198 73-273-73-273-198-198-273-73zm768 544q0 209-103 385.5t-279.5
-    279.5-385.5 103-385.5-103-279.5-279.5-103-385.5 103-385.5 279.5-279.5
-    385.5-103 385.5 103 279.5 279.5 103 385.5z"
-  />`,
-  options: {
-    viewBox: {
-      x: 1792,
-      y: 1792,
-    },
-  },
-};
