+++ src/tools/annotation/TextMarkerTool.js
@@ -8,6 +8,7 @@ import {
   removeToolState,
   getToolState,
 } from './../../stateManagement/toolState.js';
+import { textMarkerCursor } from '../cursors/index.js';
 
 /**
  * @public
@@ -233,22 +234,3 @@ const changeTextCallback = (data, eventData, doneChangingTextCallback) => {
  * @param {Object} data
  * @param {string} text - The new text
  */
-
-const textMarkerCursor = {
-  svgGroupString: `<path fill="ACTIVE_COLOR" d="M789 559l-170 450q33 0 136.5 2t160.5 2q19 0
-    57-2-87-253-184-452zm-725 1105l2-79q23-7 56-12.5t57-10.5 49.5-14.5 44.5-29
-    31-50.5l237-616 280-724h128q8 14 11 21l205 480q33 78 106 257.5t114 274.5q15
-    34 58 144.5t72 168.5q20 45 35 57 19 15 88 29.5t84 20.5q6 38 6 57 0 5-.5
-    13.5t-.5 12.5q-63 0-190-8t-191-8q-76 0-215 7t-178 8q0-43 4-78l131-28q1 0
-    12.5-2.5t15.5-3.5 14.5-4.5 15-6.5 11-8 9-11
-    2.5-14q0-16-31-96.5t-72-177.5-42-100l-450-2q-26 58-76.5 195.5t-50.5 162.5q0
-    22 14 37.5t43.5 24.5 48.5 13.5 57 8.5 41 4q1 19 1 58 0 9-2 27-58
-    0-174.5-10t-174.5-10q-8 0-26.5 4t-21.5 4q-80 14-188 14z"
-  />`,
-  options: {
-    viewBox: {
-      x: 1792,
-      y: 1792,
-    },
-  },
-};
